Reorganize factory patterns, clean up
Note, the game won't allow you to select pieces anymore.  Need to
fix this so we can work on moving them around.

Might need to build the game logic engine to keep it separated
from the graphics containers.  It seems to be getting coupled in there

May make sense to have an actual "gameContainer" which holds the:
 * skin
 * gameState
 * board
 * presentationContainer
 * rulesEngine

May make sense to break out the event loop more, and only have it call
out to the game loop on a game event, such as a GameEnd or Move.

May make sense, once moving works to start logging moves however.

May make sense to abstract game logging out so we can experiment with
different logging types (text, db, etc).